Apply an exception in the delete command

+++ HR.LeaveManagement.Application/Features/LeaveRequess/Handlers/Commands/DeleteLeaveRequessCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using HR.LeaveManagement.Application.Features.LeaveTypes.Requests.Commands;
+﻿using HR.LeaveManagement.Application.Exceptions;
+using HR.LeaveManagement.Application.Features.LeaveTypes.Requests.Commands;
 using HR.LeaveManagement.Application.Persistence.Contracts;
 using MediatR;
 
@@ -16,6 +17,10 @@ namespace HR.LeaveManagement.Application.Features.LeaveRequess.Handlers.Commands
         public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
         {
             var leaveRequest = await _leaveRequestRepository.GetAsync(request.Id);
+
+            if (leaveRequest is null)
+                throw new NotFoundException(nameof(leaveRequest), request.Id);
+
             await _leaveRequestRepository.DeleteAsync(leaveRequest);
 
             return Unit.Value;
